refactor: move ScopeHoistingFinalizer utils to mod.rs
