+++ crates/rolldown/src/module_finalizers/scope_hoisting/mod.rs
@@ -1,16 +1,22 @@
 use oxc::{
-  allocator::{Allocator, IntoIn},
+  allocator::{self, Allocator, IntoIn},
   ast::{
-    ast::{self, BindingIdentifier, Expression, IdentifierReference, MemberExpression, Statement},
+    ast::{
+      self, BindingIdentifier, ClassElement, Expression, IdentifierReference, ImportExpression,
+      MemberExpression, Statement, VariableDeclarationKind,
+    },
     Comment, NONE,
   },
   semantic::SymbolId,
   span::{Atom, GetSpan, SPAN},
 };
 use rolldown_common::{
-  AstScopes, ImportRecordIdx, ImportRecordMeta, Module, OutputFormat, Platform, SymbolRef, WrapKind,
+  AstScopes, ExportsKind, ImportRecordIdx, ImportRecordMeta, Module, ModuleType, OutputFormat,
+  Platform, SymbolRef, WrapKind,
 };
-use rolldown_ecmascript_utils::{AstSnippet, BindingPatternExt, ExpressionExt, TakeIn};
+use rolldown_ecmascript_utils::{
+  AllocatorExt, AstSnippet, BindingPatternExt, ExpressionExt, StatementExt, TakeIn,
+};
 
 mod finalizer_context;
 mod impl_visit_mut;
@@ -19,6 +25,8 @@ use rolldown_rstr::Rstr;
 use rolldown_std_utils::PathExt;
 use rolldown_utils::ecmascript::is_validate_identifier_name;
 use sugar_path::SugarPath;
+
+use crate::utils::call_expression_ext::CallExpressionExt;
 
 mod rename;
 
@@ -550,5 +558,482 @@ impl<'me, 'ast> ScopeHoistingFinalizer<'me, 'ast> {
     let original_name = symbol_ref.name(self.ctx.symbol_db);
     let canonical_name = self.canonical_name_for(symbol_ref);
     (original_name != canonical_name.as_str()).then_some((symbol_id, original_name, canonical_name))
+  }
+
+  /// rewrite toplevel `class ClassName {}` to `var ClassName = class {}`
+  fn get_transformed_class_decl(
+    &mut self,
+    class: &mut allocator::Box<'ast, ast::Class<'ast>>,
+  ) -> Option<ast::Declaration<'ast>> {
+    let scope_id = class.scope_id.get()?;
+
+    if self.scope.get_parent_id(scope_id) != Some(self.scope.root_scope_id()) {
+      return None;
+    };
+
+    let id = class.id.take()?;
+
+    if let Some(symbol_id) = id.symbol_id.get() {
+      if self.ctx.module.self_referenced_class_decl_symbol_ids.contains(&symbol_id) {
+        // class T { static a = new T(); }
+        // needs to rewrite to `var T = class T { static a = new T(); }`
+        let mut id = id.clone();
+        let new_name = self.canonical_name_for((self.ctx.id, symbol_id).into());
+        id.name = self.snippet.atom(new_name);
+        class.id = Some(id);
+      }
+    }
+    Some(self.snippet.builder.declaration_variable(
+      SPAN,
+      VariableDeclarationKind::Var,
+      self.snippet.builder.vec1(self.snippet.builder.variable_declarator(
+        SPAN,
+        VariableDeclarationKind::Var,
+        self.snippet.builder.binding_pattern(
+          ast::BindingPatternKind::BindingIdentifier(self.snippet.builder.alloc(id)),
+          NONE,
+          false,
+        ),
+        Some(Expression::ClassExpression(class.take_in(self.alloc))),
+        false,
+      )),
+      false,
+    ))
+  }
+
+  #[allow(clippy::too_many_lines, clippy::collapsible_else_if)]
+  fn try_rewrite_global_require_call(
+    &mut self,
+    call_expr: &mut ast::CallExpression<'ast>,
+  ) -> Option<Expression<'ast>> {
+    if call_expr.is_global_require_call(
+      self.scope,
+      self.ctx.symbol_db.this_method_should_be_removed_get_symbol_table(self.ctx.id),
+    ) && !call_expr.span.is_unspanned()
+    {
+      //  `require` calls that can't be recognized by rolldown are ignored in scanning, so they were not stored in `NormalModule#imports`.
+      //  we just keep these `require` calls as it is
+      if let Some(rec_id) = self.ctx.module.imports.get(&call_expr.span).copied() {
+        let rec = &self.ctx.module.import_records[rec_id];
+        // use `__require` instead of `require`
+        if rec.meta.contains(ImportRecordMeta::CALL_RUNTIME_REQUIRE) {
+          *call_expr.callee.get_inner_expression_mut() =
+            self.finalized_expr_for_symbol_ref(self.canonical_ref_for_runtime("__require"), false);
+        }
+        let rewrite_ast = match &self.ctx.modules[rec.resolved_module] {
+          Module::Normal(importee) => {
+            match importee.module_type {
+              ModuleType::Json => {
+                // Nodejs treats json files as an esm module with a default export and rolldown follows this behavior.
+                // And to make sure the runtime behavior is correct, we need to rewrite `require('xxx.json')` to `require('xxx.json').default` to align with the runtime behavior of nodejs.
+
+                // Rewrite `require(...)` to `require_xxx(...)` or `(init_xxx(), __toCommonJS(xxx_exports).default)`
+                let importee_linking_info = &self.ctx.linking_infos[importee.idx];
+                let wrap_ref_name =
+                  self.canonical_name_for(importee_linking_info.wrapper_ref.unwrap());
+                if matches!(importee.exports_kind, ExportsKind::CommonJs) {
+                  Some(self.snippet.call_expr_expr(wrap_ref_name))
+                } else {
+                  let ns_name = self.canonical_name_for(importee.namespace_object_ref);
+                  let to_commonjs_ref_name = self.canonical_name_for_runtime("__toCommonJS");
+                  Some(
+                    self.snippet.seq2_in_paren_expr(
+                      self.snippet.call_expr_expr(wrap_ref_name),
+                      ast::Expression::StaticMemberExpression(
+                        ast::StaticMemberExpression {
+                          object: self
+                            .snippet
+                            .call_expr_with_arg_expr(to_commonjs_ref_name, ns_name),
+                          property: self.snippet.id_name("default", SPAN),
+                          ..TakeIn::dummy(self.alloc)
+                        }
+                        .into_in(self.alloc),
+                      ),
+                    ),
+                  )
+                }
+              }
+              _ => {
+                // Rewrite `require(...)` to `require_xxx(...)` or `(init_xxx(), __toCommonJS(xxx_exports))`
+                let importee_linking_info = &self.ctx.linking_infos[importee.idx];
+
+                // `init_xxx`
+                let wrap_ref_expr = self
+                  .finalized_expr_for_symbol_ref(importee_linking_info.wrapper_ref.unwrap(), false);
+                if matches!(importee.exports_kind, ExportsKind::CommonJs) {
+                  // `init_xxx()`
+                  Some(ast::Expression::CallExpression(self.snippet.builder.alloc_call_expression(
+                    SPAN,
+                    wrap_ref_expr,
+                    NONE,
+                    self.snippet.builder.vec(),
+                    false,
+                  )))
+                } else {
+                  if rec.meta.contains(ImportRecordMeta::IS_REQUIRE_UNUSED) {
+                    // `init_xxx()`
+                    Some(ast::Expression::CallExpression(
+                      self.snippet.builder.alloc_call_expression(
+                        SPAN,
+                        wrap_ref_expr,
+                        NONE,
+                        self.snippet.builder.vec(),
+                        false,
+                      ),
+                    ))
+                  } else {
+                    // `xxx_exports`
+                    let namespace_object_ref_expr =
+                      self.finalized_expr_for_symbol_ref(importee.namespace_object_ref, false);
+                    let to_commonjs_ref = self.canonical_ref_for_runtime("__toCommonJS");
+                    // `__toCommonJS`
+                    let to_commonjs_expr =
+                      self.finalized_expr_for_symbol_ref(to_commonjs_ref, false);
+
+                    // `init_xxx()`
+                    let wrap_ref_call_expr =
+                      ast::Expression::CallExpression(self.snippet.builder.alloc_call_expression(
+                        SPAN,
+                        wrap_ref_expr,
+                        NONE,
+                        self.snippet.builder.vec(),
+                        false,
+                      ));
+
+                    // `__toCommonJS(xxx_exports)`
+                    let to_commonjs_call_expr =
+                      ast::Expression::CallExpression(self.snippet.builder.alloc_call_expression(
+                        SPAN,
+                        to_commonjs_expr,
+                        NONE,
+                        self.snippet.builder.vec1(ast::Argument::from(namespace_object_ref_expr)),
+                        false,
+                      ));
+
+                    // `(init_xxx(), __toCommonJS(xxx_exports))`
+                    Some(self.snippet.seq2_in_paren_expr(wrap_ref_call_expr, to_commonjs_call_expr))
+                  }
+                }
+              }
+            }
+          }
+          Module::External(importee) => {
+            let request_path =
+              call_expr.arguments.get_mut(0).expect("require should have an argument");
+
+            // Rewrite `require('xxx')` to `require('fs')`, if there is an alias that maps 'xxx' to 'fs'
+            *request_path = ast::Argument::StringLiteral(
+              self.snippet.alloc_string_literal(&importee.name, request_path.span()),
+            );
+            None
+          }
+        };
+        return rewrite_ast;
+      }
+    }
+    None
+  }
+
+  fn try_rewrite_inline_dynamic_import_expr(
+    &mut self,
+    import_expr: &mut ImportExpression<'ast>,
+  ) -> Option<Expression<'ast>> {
+    if self.ctx.options.inline_dynamic_imports {
+      let rec_id = self.ctx.module.imports.get(&import_expr.span)?;
+      let rec = &self.ctx.module.import_records[*rec_id];
+      let importee_id = rec.resolved_module;
+      match &self.ctx.modules[importee_id] {
+        Module::Normal(importee) => {
+          let importee_linking_info = &self.ctx.linking_infos[importee_id];
+          let new_expr = match importee_linking_info.wrap_kind {
+            WrapKind::Esm => {
+              // Rewrite `import('./foo.mjs')` to `(init_foo(), foo_exports)`
+              let importee_linking_info = &self.ctx.linking_infos[importee_id];
+
+              // `init_foo`
+              let importee_wrapper_ref_name =
+                self.canonical_name_for(importee_linking_info.wrapper_ref.unwrap());
+
+              // `foo_exports`
+              let importee_namespace_name = self.canonical_name_for(importee.namespace_object_ref);
+
+              // `(init_foo(), foo_exports)`
+              Some(self.snippet.promise_resolve_then_call_expr(
+                import_expr.span,
+                self.snippet.builder.vec1(self.snippet.return_stmt(
+                  self.snippet.seq2_in_paren_expr(
+                    self.snippet.call_expr_expr(importee_wrapper_ref_name),
+                    self.snippet.id_ref_expr(importee_namespace_name, SPAN),
+                  ),
+                )),
+              ))
+            }
+            WrapKind::Cjs => {
+              //  `__toESM(require_foo())`
+              let to_esm_fn_name = self.canonical_name_for_runtime("__toESM");
+              let importee_wrapper_ref_name =
+                self.canonical_name_for(importee_linking_info.wrapper_ref.unwrap());
+
+              Some(self.snippet.promise_resolve_then_call_expr(
+                import_expr.span,
+                self.snippet.builder.vec1(self.snippet.return_stmt(
+                  self.snippet.to_esm_call_with_interop(
+                    to_esm_fn_name,
+                    self.snippet.call_expr_expr(importee_wrapper_ref_name),
+                    importee.interop(),
+                  ),
+                )),
+              ))
+            }
+            WrapKind::None => {
+              // The nature of `import()` is to load the module dynamically/lazily, so imported modules would
+              // must be wrapped, so we could make sure the module is executed lazily.
+              if cfg!(debug_assertions) {
+                unreachable!()
+              }
+              None
+            }
+          };
+          return new_expr;
+        }
+        Module::External(_) => {
+          // iife format doesn't support external module
+        }
+      }
+    }
+    if matches!(self.ctx.options.format, OutputFormat::Cjs) {
+      // Convert `import('./foo.mjs')` to `Promise.resolve().then(function() { return require('foo.mjs') })`
+      let rec_id = self.ctx.module.imports.get(&import_expr.span)?;
+      let rec = &self.ctx.module.import_records[*rec_id];
+      let importee_id = rec.resolved_module;
+      match &self.ctx.modules[importee_id] {
+        Module::Normal(_importee) => {
+          let importer_chunk_id = self.ctx.chunk_graph.module_to_chunk[self.ctx.module.idx]
+            .expect("Normal module should belong to a chunk");
+          let importer_chunk = &self.ctx.chunk_graph.chunk_table[importer_chunk_id];
+          let importee_chunk_id = self.ctx.chunk_graph.entry_module_to_entry_chunk[&importee_id];
+          let importee_chunk = &self.ctx.chunk_graph.chunk_table[importee_chunk_id];
+          let import_path = importer_chunk.import_path_for(importee_chunk);
+          let new_expr = self.snippet.promise_resolve_then_call_expr(
+            import_expr.span,
+            self.snippet.builder.vec1(ast::Statement::ReturnStatement(
+              self.snippet.builder.alloc_return_statement(
+                SPAN,
+                Some(ast::Expression::CallExpression(self.snippet.builder.alloc_call_expression(
+                  SPAN,
+                  self.snippet.builder.expression_identifier_reference(SPAN, "require"),
+                  NONE,
+                  self.snippet.builder.vec1(ast::Argument::StringLiteral(
+                    self.snippet.alloc_string_literal(&import_path, import_expr.span),
+                  )),
+                  false,
+                ))),
+              ),
+            )),
+          );
+          return Some(new_expr);
+        }
+        Module::External(_) => {
+          // For `import('external')`, we just keep it as it is to preserve user's intention
+        }
+      }
+    }
+    None
+  }
+
+  #[allow(clippy::too_many_lines)]
+  fn remove_unused_top_level_stmt(&mut self, program: &mut ast::Program<'ast>) {
+    let old_body = self.alloc.take(&mut program.body);
+
+    // the first statement info is the namespace variable declaration
+    // skip first statement info to make sure `program.body` has same index as `stmt_infos`
+    old_body.into_iter().enumerate().zip(self.ctx.module.stmt_infos.iter().skip(1)).for_each(
+      |((_top_stmt_idx, mut top_stmt), stmt_info)| {
+        debug_assert!(matches!(stmt_info.stmt_idx, Some(_top_stmt_idx)));
+        if !stmt_info.is_included {
+          return;
+        }
+
+        if let Some(import_decl) = top_stmt.as_import_declaration() {
+          let rec_id = self.ctx.module.imports[&import_decl.span];
+          if self.transform_or_remove_import_export_stmt(&mut top_stmt, rec_id) {
+            return;
+          }
+        } else if let Some(export_all_decl) = top_stmt.as_export_all_declaration() {
+          let rec_id = self.ctx.module.imports[&export_all_decl.span];
+          // "export * as ns from 'path'"
+          if let Some(_alias) = &export_all_decl.exported {
+            if self.transform_or_remove_import_export_stmt(&mut top_stmt, rec_id) {
+              return;
+            }
+          } else {
+            // "export * from 'path'"
+            let rec = &self.ctx.module.import_records[rec_id];
+            match &self.ctx.modules[rec.resolved_module] {
+              Module::Normal(importee) => {
+                let importee_linking_info = &self.ctx.linking_infos[importee.idx];
+                if matches!(importee_linking_info.wrap_kind, WrapKind::Esm) {
+                  let wrapper_ref_name =
+                    self.canonical_name_for(importee_linking_info.wrapper_ref.unwrap());
+                  program.body.push(self.snippet.call_expr_stmt(wrapper_ref_name));
+                }
+
+                match importee.exports_kind {
+                  ExportsKind::Esm => {
+                    if importee_linking_info.has_dynamic_exports {
+                      let re_export_fn_name = self.canonical_name_for_runtime("__reExport");
+                      let importer_namespace_name =
+                        self.canonical_name_for(self.ctx.module.namespace_object_ref);
+                      // __reExport(exports, otherExports)
+                      let importee_namespace_name =
+                        self.canonical_name_for(importee.namespace_object_ref);
+                      program.body.push(
+                        self
+                          .snippet
+                          .call_expr_with_2arg_expr(
+                            re_export_fn_name,
+                            importer_namespace_name,
+                            importee_namespace_name,
+                          )
+                          .into_in(self.alloc),
+                      );
+                    }
+                  }
+                  ExportsKind::CommonJs => {
+                    let re_export_fn_name = self.canonical_name_for_runtime("__reExport");
+                    let importer_namespace_name =
+                      self.canonical_name_for(self.ctx.module.namespace_object_ref);
+                    // __reExport(exports, __toESM(require_xxxx()))
+                    let to_esm_fn_name = self.canonical_name_for_runtime("__toESM");
+                    let importee_wrapper_ref_name =
+                      self.canonical_name_for(importee_linking_info.wrapper_ref.unwrap());
+                    program.body.push(
+                      self
+                        .snippet
+                        .alloc_call_expr_with_2arg_expr_expr(
+                          re_export_fn_name,
+                          self.snippet.id_ref_expr(importer_namespace_name, SPAN),
+                          self.snippet.to_esm_call_with_interop(
+                            to_esm_fn_name,
+                            self.snippet.call_expr_expr(importee_wrapper_ref_name),
+                            importee.interop(),
+                          ),
+                        )
+                        .into_in(self.alloc),
+                    );
+                  }
+                  ExportsKind::None => {}
+                }
+              }
+              Module::External(_importee) => {
+                match self.ctx.options.format {
+                  rolldown_common::OutputFormat::Esm
+                  | rolldown_common::OutputFormat::Iife
+                  | rolldown_common::OutputFormat::Umd
+                  | rolldown_common::OutputFormat::Cjs => {
+                    // Just remove the statement
+                    return;
+                  }
+                  rolldown_common::OutputFormat::App => {
+                    unreachable!()
+                  }
+                }
+              }
+            }
+
+            return;
+          }
+        } else if let Some(default_decl) = top_stmt.as_export_default_declaration_mut() {
+          use ast::ExportDefaultDeclarationKind;
+          match &mut default_decl.declaration {
+            decl @ ast::match_expression!(ExportDefaultDeclarationKind) => {
+              let expr = decl.to_expression_mut();
+              // "export default foo;" => "var default = foo;"
+              let canonical_name_for_default_export_ref =
+                self.canonical_name_for(self.ctx.module.default_export_ref);
+              top_stmt = self
+                .snippet
+                .var_decl_stmt(canonical_name_for_default_export_ref, expr.take_in(self.alloc));
+            }
+            ast::ExportDefaultDeclarationKind::FunctionDeclaration(func) => {
+              // "export default function() {}" => "function default() {}"
+              // "export default function foo() {}" => "function foo() {}"
+              if func.id.is_none() {
+                let canonical_name_for_default_export_ref =
+                  self.canonical_name_for(self.ctx.module.default_export_ref);
+                func.id = Some(self.snippet.id(canonical_name_for_default_export_ref, SPAN));
+              }
+              top_stmt = ast::Statement::FunctionDeclaration(func.take_in(self.alloc));
+            }
+            ast::ExportDefaultDeclarationKind::ClassDeclaration(class) => {
+              // "export default class {}" => "class default {}"
+              // "export default class Foo {}" => "class Foo {}"
+              if class.id.is_none() {
+                let canonical_name_for_default_export_ref =
+                  self.canonical_name_for(self.ctx.module.default_export_ref);
+                class.id = Some(self.snippet.id(canonical_name_for_default_export_ref, SPAN));
+              }
+              top_stmt = ast::Statement::ClassDeclaration(class.take_in(self.alloc));
+            }
+            _ => {}
+          }
+        } else if let Some(named_decl) = top_stmt.as_export_named_declaration_mut() {
+          if named_decl.source.is_none() {
+            if let Some(decl) = &mut named_decl.declaration {
+              // `export var foo = 1` => `var foo = 1`
+              // `export function foo() {}` => `function foo() {}`
+              // `export class Foo {}` => `class Foo {}`
+              top_stmt = ast::Statement::from(decl.take_in(self.alloc));
+            } else {
+              // `export { foo }`
+              // Remove this statement by ignoring it
+              return;
+            }
+          } else {
+            // `export { foo } from 'path'`
+            let rec_id = self.ctx.module.imports[&named_decl.span];
+            if self.transform_or_remove_import_export_stmt(&mut top_stmt, rec_id) {
+              return;
+            }
+          }
+        }
+
+        program.body.push(top_stmt);
+      },
+    );
+  }
+
+  fn process_fn(
+    &mut self,
+    symbol_binding_id: Option<&BindingIdentifier<'ast>>,
+    name_binding_id: Option<&BindingIdentifier<'ast>>,
+  ) -> Option<()> {
+    if !self.ctx.options.keep_names {
+      return None;
+    }
+    let (_, original_name, _) = self.get_conflicted_info(name_binding_id.as_ref()?)?;
+    let (symbol_id, _, canonical_name) = self.get_conflicted_info(symbol_binding_id.as_ref()?)?;
+    let original_name: Rstr = original_name.into();
+    let new_name = canonical_name.clone();
+    let insert_position = self.ctx.cur_stmt_index + 1;
+    self.ctx.keep_name_statement_to_insert.push((
+      insert_position,
+      symbol_id,
+      original_name,
+      new_name,
+    ));
+    None
+  }
+
+  fn keep_name_helper_for_class(
+    &mut self,
+    id: Option<&BindingIdentifier<'ast>>,
+  ) -> Option<ClassElement<'ast>> {
+    if !self.ctx.options.keep_names {
+      return None;
+    }
+    let (_, original_name, _) = self.get_conflicted_info(id.as_ref()?)?;
+    let original_name: Rstr = original_name.into();
+    Some(self.snippet.static_block_keep_name_helper(&original_name))
   }
 }
